+++ tests/test_integration.py
@@ -7,12 +7,14 @@ from tests.testcase import TestCase
 
 __author__ = "Alex Laird"
 __copyright__ = "Copyright 2024, Alex Laird"
-__version__ = "1.0.1"
+__version__ = "1.0.2"
 
 
 @unittest.skipIf(os.environ.get("INTEGRATION_TESTS", "False") != "True",
                  "Skipping, INTEGRATION_TESTS=True was not set in the environment")
 class TestIntegration(TestCase):
+    amazon_session = None
+
     @classmethod
     def setUpClass(cls):
         cls.credentials_found = os.environ.get("AMAZON_USERNAME") and os.environ.get("AMAZON_PASSWORD")