+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## [Unreleased](https://github.com/alexdlaird/amazon-orders-python/compare/1.0.1...HEAD)
 ### Added
 - `IODefault` for I/O operations, which can be extended to use something other than `print()` and `input()`.
+- Documentation improvements.
+- Test improvements.
 
 ### Removed
 `Orders.print_output` variable, `cli.py` now handles output.