+++ CHANGELOG.md
@@ -3,7 +3,9 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/alexdlaird/amazon-orders-python/compare/1.0.1...HEAD)
+## [Unreleased](https://github.com/alexdlaird/amazon-orders-python/compare/1.0.2...HEAD)
+
+## [1.0.2](https://github.com/alexdlaird/pyngrok/compare/1.0.1...1.0.2) - 2024-01-18
 ### Added
 - `IODefault` for I/O operations, which can be extended to use something other than `print()` and `input()`.
 - Documentation improvements.